+++ src/kernel/vid/print.c
@@ -96,7 +96,7 @@ void VidpFlush(void) {
         memcpy(
             VidpBackBuffer + (ptrdiff_t)((VidpFlushY + WrapLine) * VidpPitch),
             VidpFrontBuffer,
-            (size_t)((VidpFlushLines - WrapLine) * VidpPitch));
+            (size_t)(VidpFlushLines - WrapLine) * VidpPitch);
     }
 
     VidpPendingFullFlush = false;
@@ -126,7 +126,7 @@ static void ScrollUp(void) {
 
         /* Then, we can either set pixel by pixel, or use memset (if we have a solid background). */
         if (VidpBackground == (VidpBackground & 0xFF) * 0x01010101) {
-            memset(Buffer, VidpBackground & 0xFF, (size_t)(VidpWidth * 4));
+            memset(Buffer, (int)(VidpBackground & 0xFF), (size_t)(VidpWidth * 4));
         } else {
             for (int j = 0; j < VidpWidth; j++) {
                 Buffer[j] = VidpBackground;
@@ -196,7 +196,7 @@ static void DrawCharacter(char Character) {
         uint16_t FrontY = (VidpCursorY + VidpRingTop + Top) % VidpHeight;
         uint32_t *Buffer = (uint32_t *)&VidpFrontBuffer[FrontY * VidpPitch + VidpCursorX * 4];
         if (VidpBackground == (VidpBackground & 0xFF) * 0x01010101) {
-            memset(Buffer, VidpBackground & 0xFF, (size_t)(LeftLimit * 4));
+            memset(Buffer, (int)(VidpBackground & 0xFF), (size_t)(LeftLimit * 4));
         } else {
             for (int j = 0; j < TopLimit; j++) {
                 Buffer[j] = VidpBackground;
