+++ src/sdk/crt/math/cosf.c
@@ -59,9 +59,11 @@ float cosf(float x) {
     uint32_t xi_abs = xi & ~0x80000000;
     if (xi_abs >= 0x7f800000) {
         if (!(xi & 0x7fffff)) {
+            /* NOLINTNEXTLINE(misc-redundant-expression) */
             __raise_errnof(0.0F / 0.0F, EDOM);
         }
 
+        /* NOLINTNEXTLINE(misc-redundant-expression) */
         return x - x;
     }
 
@@ -74,7 +76,7 @@ float cosf(float x) {
     /* The core approx function works on the range [-pi/4;pi/4], if we're inside it, no need
      * to reduce the argument. */
     if (xi_abs < 0x3f490fdb) {
-        return __cosf(x);
+        return (float)__cosf(x);
     }
 
     /* Otherwise, reduce and use the quadrant to choose the right operation to execute. */
@@ -82,13 +84,13 @@ float cosf(float x) {
     double res = __rem_pio2f(x, &q);
     switch (q & 3) {
         case 0:
-            return __cosf(res);
+            return (float)__cosf(res);
         case 1:
-            return -__sinf(res);
+            return (float)-__sinf(res);
         case 2:
-            return -__cosf(res);
+            return (float)-__cosf(res);
         case 3:
-            return __sinf(res);
+            return (float)__sinf(res);
         default:
             unreachable();
     }
