+++ src/sdk/crt/stdio/vsnprintf.c
@@ -31,7 +31,7 @@ static void put_buf(const void *buffer, int size, void *context) {
     context_t *str_context = context;
 
     if ((intmax_t)str_context->bufsz < size) {
-        size = str_context->bufsz;
+        size = (int)str_context->bufsz;
     }
 
     if (size > 0) {
