+++ src/sdk/crt/stdlib/strtoull.c
@@ -51,7 +51,7 @@ unsigned long long strtoull(const char *CRT_RESTRICT nptr, char **CRT_RESTRICT e
 
     while (true) {
         unsigned long long last = value;
-        int digit = *nptr;
+        int digit = (int)*nptr;
 
         /* First two cases are aA-fF, offset by +0x0A. */
         if (islower(digit)) {
